@@ -19,7 +19,7 @@ name = "idp_common"
 version = "0.3.13"
 description = "Common utilities for GenAI IDP Accelerator patterns"
 authors = [{ name = "AWS", email = "noreply@amazon.com" }]
-requires-python = ">=3.9,<3.14"
+requires-python = ">=3.10,<3.14"
 dependencies = [
   "boto3==1.39.7",      # Core dependency for AWS services
   "jsonschema>=4.25.1",
@@ -98,6 +98,17 @@ agents = [
   "strands-agents>=1.0.0; python_version>='3.10'",
   "strands-agents-tools>=0.2.2; python_version>='3.10'",
   "bedrock-agentcore>=0.1.1; python_version>='3.10'",    # Specifically for the code interpreter tool
+]
+
+# Code intelligence module dependencies
+code_intel = [
+    "requests==2.32.4",
+    "strands-agents>=1.0.0",
+    "strands-agents-tools>=0.2.2",
+    "bedrock-agentcore>=0.1.1",
+    "PyYAML>=6.0.0",
+    "pathspec>=0.11.0",
+    "chardet>=5.0.0"
 ]
 
 # Document service factory dependencies (includes both appsync and dynamodb support)
