@@ -296,8 +296,16 @@ def create_analytics_agent(
         get_table_info,  # Detailed schema for specific tables
     ]
 
-    # Get model ID from environment variable
-    model_id = os.environ.get("DOCUMENT_ANALYSIS_AGENT_MODEL_ID")
+    # Get model ID using configuration helper (checks env vars, config table, then defaults)
+    try:
+        from ...config import ConfigurationManager
+        from .config import get_analytics_model_id
+
+        config_manager = ConfigurationManager()
+        model_id = get_analytics_model_id(config_manager=config_manager)
+    except Exception as e:
+        logger.warning(f"Failed to get analytics model ID, using default: {e}")
+        model_id = "us.anthropic.claude-3-7-sonnet-20250219-v1:0"
 
     bedrock_model = create_strands_bedrock_model(
         model_id=model_id, boto_session=session
