@@ -10,14 +10,14 @@ appropriate agent based on the user's request and the capabilities of each agent
 
 import logging
 import os
-import types
-from typing import Any, Dict, List
+from typing import Any, AsyncIterator, Dict, List
 
+import boto3
 import strands
 from strands import tool
 
-from ..common.config import load_result_format_description
 from ..common.strands_bedrock_model import create_strands_bedrock_model
+from .config import get_chat_companion_model_id
 
 logger = logging.getLogger(__name__)
 
@@ -44,53 +44,124 @@ def create_orchestrator_agent(
     # Create tool functions for each specialized agent
     tools = []
 
-    # Extract monitoring context to pass to sub-agents
-    job_id = kwargs.get("job_id")
-    user_id = kwargs.get("user_id")
+    logger.info("Creating orchestrator with sub-agent streaming enabled")
 
     for agent_id in agent_ids:
         # Create a unique function name based on the agent ID
         func_name = f"{agent_id.replace('-', '_')}_agent"
 
         def create_tool_function(aid, fname):
-            def tool_func(query: str) -> str:
+            # Async streaming version - always enabled
+            async def tool_func(query: str) -> AsyncIterator:
+                """Route query to specialized agent with streaming."""
+                import asyncio
+
                 try:
                     from ..factory import agent_factory
 
+                    # Create fresh boto3 session for each sub-agent to isolate connection pools
+                    # This prevents connection pool conflicts between concurrent sub-agents
+                    sub_session = boto3.Session()
+
+                    sub_kwargs = {k: v for k, v in kwargs.items() if k != "session_id"}
                     specialized_agent = agent_factory.create_agent(
                         agent_id=aid,
                         config=config,
-                        session=session,
-                        job_id=job_id,
-                        user_id=user_id,
-                        **{
-                            k: v
-                            for k, v in kwargs.items()
-                            if k not in ["job_id", "user_id"]
-                        },
+                        session=sub_session,  # Use fresh session per sub-agent
+                        **sub_kwargs,
                     )
 
-                    # Use context manager to properly handle MCP clients
+                    # Stream sub-agent events with timeout per event
+                    timeout_seconds = int(
+                        os.environ.get("SUBAGENT_TIMEOUT_SECONDS", "120")
+                    )
+                    logger.info(
+                        f"Starting sub-agent {aid} with {timeout_seconds}s timeout"
+                    )
+
                     with specialized_agent:
-                        response = specialized_agent(query)
-                    return str(response)
+                        stream = specialized_agent.stream_async(query)
+                        last_event_time = asyncio.get_event_loop().time()
+
+                        try:
+                            async for event in stream:
+                                # Check if we've exceeded timeout since last event
+                                current_time = asyncio.get_event_loop().time()
+                                if current_time - last_event_time > timeout_seconds:
+                                    error_msg = f"Sub-agent {aid} timed out (no events for {timeout_seconds}s)"
+                                    logger.error(error_msg)
+                                    yield f"Error: {error_msg}"
+                                    break
+
+                                last_event_time = current_time
+
+                                # Log what events we're receiving from sub-agent
+                                logger.info(
+                                    f"Sub-agent {aid} event: {list(event.keys())}"
+                                )
+
+                                # Yield each event for the orchestrator to process
+                                yield event
+
+                                # When we get the final result, yield it as the tool return
+                                if "result" in event:
+                                    result = event["result"]
+                                    response_text = str(result)
+
+                                    # Extract JSON from markdown code blocks if present
+                                    from ..common.response_utils import (
+                                        extract_json_from_markdown,
+                                    )
+
+                                    cleaned_response = extract_json_from_markdown(
+                                        response_text
+                                    )
+
+                                    # Check if this is structured data (table/plot)
+                                    try:
+                                        import json
+
+                                        parsed = json.loads(cleaned_response)
+                                        response_type = parsed.get("responseType")
+
+                                        if response_type in ["table", "plotData"]:
+                                            # Yield a special event to signal structured data
+                                            yield {
+                                                "structured_data_detected": True,
+                                                "responseType": response_type,
+                                            }
+                                            logger.info(
+                                                f"Sub-agent {aid} returning {response_type} data"
+                                            )
+                                    except (
+                                        json.JSONDecodeError,
+                                        AttributeError,
+                                        TypeError,
+                                    ):
+                                        # Not JSON or doesn't have responseType - continue normally
+                                        pass
+
+                                    logger.info(
+                                        f"Sub-agent {aid} completed, yielding final result"
+                                    )
+                                    yield cleaned_response
+                                    break
+                        except asyncio.TimeoutError:
+                            error_msg = f"Sub-agent {aid} timed out after {timeout_seconds} seconds"
+                            logger.error(error_msg)
+                            yield f"Error: {error_msg}. The agent took too long to respond."
 
                 except Exception as e:
-                    logger.error(f"Error in {aid}: {e}")
-                    return f"Error processing query with {aid}: {str(e)}"
+                    logger.error(f"Error in sub-agent {aid}: {e}", exc_info=True)
+                    yield f"Error processing query with {aid}: {str(e)}"
 
-            new_func = types.FunctionType(
-                tool_func.__code__,
-                tool_func.__globals__,
-                fname,
-                tool_func.__defaults__,
-                tool_func.__closure__,
-            )
-
-            new_func.__doc__ = f"Route query to agent {aid}"
+            # Rename the function to match the desired tool name
+            tool_func.__name__ = fname
+            tool_func.__qualname__ = fname
             # Store the original agent_id as a custom attribute
-            new_func._original_agent_id = aid
-            return tool(new_func)
+            tool_func._original_agent_id = aid
+            # Apply tool decorator and return
+            return tool(tool_func)
 
         # Create the tool function for this specific agent
         tool_func = create_tool_function(agent_id, func_name)
@@ -124,41 +195,98 @@ def create_orchestrator_agent(
     # Create system prompt with agent descriptions
     agent_names_description_sample_queries = "\n".join(agent_descriptions)
 
-    system_prompt = f"""You are an intelligent orchestrator that routes user queries to specialized agents.
+    system_prompt = f"""You are IDP Companion, an intelligent AI Assistant that can answer question about the IDP app. 
+    Specifically, as the main agent or the orchestrator, you cooridnate and leverage specialized agents to answer user queries.
 
-Available specialized agents:
+# Available Specialized Agents
 {agent_names_description_sample_queries}
 
-# Role
-Your role is to:
-1. Analyze the user's query to understand what they're asking for
-2. Select the most appropriate specialized agent based on the query content and agent capabilities
-3. Route the query to that agent using the corresponding tool
-4. Return the agent's response to the user in the result format specified below:
+# Your Workflow
+1. **Analyze** the user's query to understand what information is needed
+2. **Select** the most appropriate specialized agent(s) based on capabilities and sample queries
+3. **Call** the agent tool(s) to gather information - you may call multiple agents if needed
+4. **Evaluate** if the response fully answers the user's question
+5. **Stop** if you have sufficient information - do NOT make redundant calls
+6. **Process** the responses and format your final answer appropriately
 
-# Result format
-Your final response needs to formatted in a specific way for the downstream user to understand it. Here is a description of the result format:
-```markdown
-{load_result_format_description()}
+# Agent Response Format
+Specialized agents return JSON responses in this structure:
+```json
+{{
+  "responseType": "text" | "table" | "plotData",
+  "content": "...",        // For text responses
+  "tableData": {{...}},    // For table responses  
+  "plotData": {{...}}      // For plot responses
+}}
 ```
 
-# Guidelines
-- Choose the agent whose description and sample queries best match the user's request
-- If a query could fit multiple agents, choose the most specific/specialized one
-- Always use exactly one agent tool per query
-- Pass the user's original query directly to the selected agent
-- Return the agent's response without modification
+# How to Format Your Final Response
 
-Remember: Your job is to route queries, not to answer them directly. Always use one of the available agent tools.
+## Text Responses (responseType: "text")
+When an agent returns text:
+- Extract and understand the `content` field
+- **Evaluate if the response fully answers the user's question**
+- If the response is complete and satisfactory, **STOP and respond immediately**
+- Only call additional agents if the first response is incomplete or missing critical information
+- Once you have sufficient information, respond in **natural conversational text** (NOT JSON)
+- If a single agent fully answers the question, you can pass through its answer directly
+- For multi-agent responses, synthesize the information into a coherent answer
+- **NEVER call the same agent twice with the same or similar query**
 
-If the agents response is a json already matching the above result format, return it directly without any modifications or additional interpretation of the results.
+Example:
+- Agent returns: `{{"responseType": "text", "content": "The repository contains 5 Python files..."}}`
+- You respond: "Based on the analysis, this repository contains 5 Python files..."
+
+## Structured Data Responses (responseType: "table" or "plotData")
+When an agent returns structured data:
+- **Return ONLY the raw JSON exactly as received**
+- Do NOT add any text, explanations, or formatting
+- Do NOT modify the JSON structure
+- The UI will handle rendering
+
+Example:
+- Agent returns: `{{"responseType": "table", "tableData": {{...}}}}`
+- You respond: `{{"responseType": "table", "tableData": {{...}}}}`
+
+# Agent Selection Guidelines
+- Match the query to the agent whose description and sample queries are most relevant
+- Prefer more specialized agents over general ones when applicable
+- You can call multiple agents sequentially to gather all needed information
+- Pass clear, specific queries to each agent
+- **IMPORTANT**: If a user query requires understanding the codebase or code structure, check if the Code Intelligence Agent is available in your tools. If it is NOT available, politely inform the user that you cannot answer code-related questions because the Code Intelligence Agent has been disabled for this session.
+
+# CRITICAL: Handling Unavailable Agents
+- **Code Intelligence Agent**: If the user asks about code, codebase structure, implementation details, or repository information, check if the Code Intelligence Agent is in your available tools
+- If the Code Intelligence Agent is NOT available and the query requires code understanding, respond with:
+  "I'm unable to answer questions about the codebase because the Code Intelligence Agent is currently disabled. You can enable it using the checkbox below the chat box to ask codebase-related questions."
+- Do NOT attempt to answer code-related questions without the Code Intelligence Agent
+- Do NOT make up information about the codebase
+
+# CRITICAL: When to Stop Calling Tools
+- **Stop immediately** if a single agent call fully answers the user's question
+- **Do NOT** make redundant calls to the same agent with similar queries
+- **Do NOT** call additional agents "just to be thorough" if you already have sufficient information
+- Only call multiple agents if the first response is incomplete or you need different types of information
+- Trust the agent responses - they are designed to be comprehensive
+
+# Key Rules
+- For text responses: Be conversational and helpful - return natural language, NOT JSON
+- For table/plot responses: Return ONLY the JSON with zero additional text
+- Synthesize information from multiple agents when needed
+- Keep responses clear and user-friendly
 """
 
-    # Get model ID from environment variable
-    model_id = os.environ.get(
-        "DOCUMENT_ANALYSIS_AGENT_MODEL_ID",
-        "us.anthropic.claude-3-7-sonnet-20250219-v1:0",
-    )
+    # Get model ID using configuration helper (checks env vars, config table, then defaults)
+    try:
+        from ...config import ConfigurationManager
+
+        config_manager = ConfigurationManager()
+        model_id = get_chat_companion_model_id(config_manager=config_manager)
+    except Exception as e:
+        logger.warning(f"Failed to get chat companion model ID, using default: {e}")
+        model_id = config.get(
+            "default_model_id", "us.anthropic.claude-3-7-sonnet-20250219-v1:0"
+        )
 
     # Create the orchestrator agent
     model = create_strands_bedrock_model(
@@ -166,10 +294,14 @@ If the agents response is a json already matching the above result format, retur
         session=session,
     )
 
+    # Get hooks from kwargs if provided
+    hooks = kwargs.get("hooks", [])
+
     orchestrator = strands.Agent(
         system_prompt=system_prompt,
         model=model,
         tools=tools,
+        hooks=hooks,  # Pass hooks during agent creation
         callback_handler=None,
     )
 
