@@ -12,11 +12,24 @@ import unittest
 from unittest.mock import patch
 
 from idp_common.assessment.service import AssessmentService
+from idp_common.config.models import IDPConfig
 from idp_common.models import Document, Page, Section, Status
 
 
 class TestAssessmentEnabledProperty(unittest.TestCase):
     """Test assessment service enabled property behavior."""
+
+    def _create_mock_image_data(self) -> bytes:
+        """Create a valid mock image data for testing."""
+        import io
+
+        from PIL import Image
+
+        # Create a simple 100x100 white image
+        img = Image.new("RGB", (100, 100), color="white")
+        img_bytes = io.BytesIO()
+        img.save(img_bytes, format="JPEG")
+        return img_bytes.getvalue()
 
     def setUp(self):
         """Set up test fixtures."""
@@ -80,7 +93,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         config["assessment"]["enabled"] = True
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         with (
             patch("idp_common.s3.get_json_content") as mock_get_json,
@@ -95,7 +109,7 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
                 "metadata": {},
             }
             mock_get_text.return_value = "Invoice #INV-12345\nAmount: $100.00"
-            mock_prepare_image.return_value = b"mock_image_data"
+            mock_prepare_image.return_value = self._create_mock_image_data()
 
             # Mock Bedrock response
             mock_invoke_model.return_value = {
@@ -134,7 +148,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         config["assessment"]["enabled"] = False
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         with (
             patch("idp_common.s3.get_json_content") as mock_get_json,
@@ -164,7 +179,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         # Explicitly do not set 'enabled' property
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         with (
             patch("idp_common.s3.get_json_content") as mock_get_json,
@@ -179,15 +195,19 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
                 "metadata": {},
             }
             mock_get_text.return_value = "Invoice #INV-12345\nAmount: $100.00"
-            mock_prepare_image.return_value = b"mock_image_data"
+            mock_prepare_image.return_value = self._create_mock_image_data()
 
             # Mock Bedrock response
             mock_invoke_model.return_value = {
-                "content": [
-                    {
-                        "text": '{"invoice_number": {"confidence": 0.95, "confidence_reason": "Clear text"}}'
+                "output": {
+                    "message": {
+                        "content": [
+                            {
+                                "text": '{"invoice_number": {"confidence": 0.95, "confidence_reason": "Clear text"}}'
+                            }
+                        ]
                     }
-                ],
+                },
                 "metering": {
                     "inputTokens": 1000,
                     "outputTokens": 200,
@@ -212,7 +232,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         config["assessment"]["enabled"] = "true"
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         with (
             patch("idp_common.s3.get_json_content") as mock_get_json,
@@ -227,7 +248,7 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
                 "metadata": {},
             }
             mock_get_text.return_value = "Invoice #INV-12345\nAmount: $100.00"
-            mock_prepare_image.return_value = b"mock_image_data"
+            mock_prepare_image.return_value = self._create_mock_image_data()
 
             # Mock Bedrock response
             mock_invoke_model.return_value = {
@@ -266,7 +287,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         config["assessment"]["enabled"] = "false"
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         with (
             patch("idp_common.s3.get_json_content") as mock_get_json,
@@ -295,7 +317,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         config = {"classes": self.base_config["classes"]}
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         with (
             patch("idp_common.s3.get_json_content") as mock_get_json,
@@ -310,7 +333,7 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
                 "metadata": {},
             }
             mock_get_text.return_value = "Invoice #INV-12345\nAmount: $100.00"
-            mock_prepare_image.return_value = b"mock_image_data"
+            mock_prepare_image.return_value = self._create_mock_image_data()
 
             # Mock Bedrock response
             mock_invoke_model.return_value = {
@@ -350,7 +373,8 @@ class TestAssessmentEnabledProperty(unittest.TestCase):
         config["assessment"]["enabled"] = False
 
         # Initialize assessment service
-        assessment_service = AssessmentService(config=config)
+        idp_config = IDPConfig.model_validate(config)
+        assessment_service = AssessmentService(config=idp_config)
 
         # Process document section
         result_document = assessment_service.process_document_section(
