@@ -1,15 +1,17 @@
 import json
 import logging
+from typing import Optional
+from idp_common.config import AssessmentConfig
 from idp_common.utils import normalize_boolean_value
 
 logger = logging.getLogger(__name__)
 
 class AssessmentValidator:
-    def __init__(self, extraction_data, assessment_config=None, enable_missing_check=False, enable_count_check=False):
+    def __init__(self, extraction_data, assessment_config: Optional[ AssessmentConfig ]=None, enable_missing_check=False, enable_count_check=False):
         self.extraction_data = extraction_data
         self.inference_result = extraction_data.get('inference_result', {})
         self.explainability_info = extraction_data.get('explainability_info', [])
-        self.assessment_config = assessment_config or {}
+        self.assessment_config = assessment_config or AssessmentConfig() 
         self.enable_missing_check = enable_missing_check
         self.enable_count_check = enable_count_check
         
@@ -107,9 +109,8 @@ class AssessmentValidator:
     def check_explainability_exists(self):
         """Check if explainability_info has at least one element when expected attributes exist"""
         expected_attributes = set(self.inference_result.keys())
-        assessment_enabled = normalize_boolean_value(self.assessment_config.get('enabled', False))
-        logger.info(f"Check Explainability Exists: {assessment_enabled}")
-        if (assessment_enabled and
+        logger.info(f"Check Explainability Exists: {self.assessment_config.enabled}")
+        if (self.assessment_config.enabled and
                 expected_attributes and
                 (not self.explainability_info or not self.explainability_info[0])):
             return {
