@@ -14,8 +14,17 @@ import useAppContext from '../contexts/app';
 import DocumentsRoutes from './DocumentsRoutes';
 import DocumentsQueryRoutes from './DocumentsQueryRoutes';
 import DocumentsAnalyticsRoutes from './DocumentsAnalyticsRoutes';
+import AgentChatRoutes from './AgentChatRoutes';
 
-import { DOCUMENTS_PATH, DEFAULT_PATH, LOGIN_PATH, LOGOUT_PATH, DOCUMENTS_KB_QUERY_PATH, DOCUMENTS_ANALYTICS_PATH } from './constants';
+import {
+  DOCUMENTS_PATH,
+  DEFAULT_PATH,
+  LOGIN_PATH,
+  LOGOUT_PATH,
+  DOCUMENTS_KB_QUERY_PATH,
+  DOCUMENTS_ANALYTICS_PATH,
+  AGENT_CHAT_PATH,
+} from './constants';
 
 const logger = new ConsoleLogger('AuthRoutes');
 
@@ -33,6 +42,7 @@ const AuthRoutes = ({ redirectParam }) => {
   return (
     <SettingsContext.Provider value={settingsContextValue}>
       <Routes>
+        <Route path={`${AGENT_CHAT_PATH}/*`} element={<AgentChatRoutes />} />
         <Route path={`${DOCUMENTS_KB_QUERY_PATH}/*`} element={<DocumentsQueryRoutes />} />
         <Route path={`${DOCUMENTS_ANALYTICS_PATH}/*`} element={<DocumentsAnalyticsRoutes />} />
         <Route path={`${DOCUMENTS_PATH}/*`} element={<DocumentsRoutes />} />
